pyproject.toml
[tool.poetry]
name = "project-social-network1-django"
version = "0.1.0"
description = ""
authors = ["Julia-Gareeva <uliastekolnikova12822@gmail.com>"]
readme = "README.md"
packages = [{include = "project_social_network1_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
